Streamlining the test run script
Most of these changes are pretty minor - just trying to streamline
the code a bit to reduce overload on students. To this end, I
simplified the parsing of HPCP a bit by using the `col_types`
argument to `read_csv`, rather than `mutate` to force it to be
read as a character.

With readr creating tbl_df objects, the `str` function is no longer
needed to see the type of each column: the default print method
works great.

For the challenges, I think it might be a good idea to not
encourage them to use ggplot within a loop like that - instead
I added some challenges that will require them to apply what they
hopefully learned about mutate, group_by, and summarize, along with
some new functions (e.g., `count` and `filter`), so that they can
check out the help files.